requirements.txt
google-generativeai
pinecone
python-dotenv
langgraph